feat: prototype for teamProject solution
- feat: introduce custom configuration option
  Update pom.xml with a better default authorization url
- feat: improve logging of jwt
- fix: add "Atlas users" as default system role
- feat: add more log statements for PermissionManager
- feat: ensure /user/me endpoint also triggers the UPDATE_TOKEN filter
- feat: ensure the teamproject is stored per user
   ...and allow reading current teamproject from cache in case of a
   request to /user/refresh endpoint
- feat: main logic in new filter class TeamProjectBasedAuthorizingFilter
- fix: ensure reset of roles always happens
- feat: remove unnecessary method from PermissionManager
- fix: use lower() in SQL query itself for finding login
- fix: take login from shiro-parsed principal instead of DB
  ... to avoid the issue where the login is all lowercase in db
- feat: move the defaultRoles definition into AtlasSecurity
- fix: move authorizationMode check to PostConstruct
  ...to avoid NullPointerException as attributes
  are not yet wired when in constructor

- fix: remove session.stop() call from UpdateAccessTokenFilter
...and therefore from the flow of endpoints like /user/refresh.
Not sure why this was added there, as the /user/logout should be
the place to remove a session.
This solves a org.apache.shiro.subject.support.DisabledSessionException.
If the worry is that logout won`t be called, then the
expiry time should just be set to a short period.
The adjustment in JwtAuthRealm.java was to deal with the side
effect that occurred after the removal of the .stop in
the UpdateAccessTokenFilter filter:
java.lang.ClassCastException: io.buji.pac4j.subject.Pac4jPrincipal cannot be cast to java.lang.String

- fix: do not create a new session when requesting current session