fix: remove session.stop() call from UpdateAccessTokenFilter
...and therefore from the flow of endpoints like /user/refresh.
Not sure why this was added there, as the /user/logout should be
the place to remove a session.
This solves a org.apache.shiro.subject.support.DisabledSessionException.

If the worry is that logout won`t be called, then the
expiry time should just be set to a short period.

The adjustment in JwtAuthRealm.java was to deal with the side
effect that occurred after the removal of the .stop in
the UpdateAccessTokenFilter filter:
java.lang.ClassCastException: io.buji.pac4j.subject.Pac4jPrincipal cannot be cast to java.lang.String